Jukka/login refactor (#132)
* Refactor password verification logic into verifyPassword()

* Refactor: Move session.js to backend/auth and update imports and test mock paths
